feat: notifications refactor
