+++ ui/portal/web/src/hooks/useNotifications.ts
@@ -1,50 +1,141 @@
-import { useStorage } from "@left-curve/store";
+import { createEventBus, useConfig, useStorage } from "@left-curve/store";
 import { format, isToday } from "date-fns";
 import { useCallback, useMemo } from "react";
-import type { Notifications } from "~/app.provider";
+import { type Client as GraphqlSubscriptionClient, createClient } from "graphql-ws";
 
-type UseNotificationsParameters = {
-  maxNotifications?: number;
+import { GRAPHQL_URI } from "../../store.config";
+
+import type { AnyCoin } from "@left-curve/store/types";
+import type { Account } from "@left-curve/dango/types";
+
+export type NotificationsMap = {
+  submit_tx:
+    | { isSubmitting: true; txResult?: never }
+    | { isSubmitting: false; txResult: { hasSucceeded: boolean; message: string } };
+  transfer: {
+    amount: number;
+    coin: AnyCoin;
+    fromAddress: string;
+    toAddress: string;
+    type: "received" | "sent";
+  };
 };
 
-export function useNotifications(parameters: UseNotificationsParameters) {
-  const { maxNotifications } = parameters;
+export type Notifications<key extends keyof NotificationsMap = keyof NotificationsMap> = {
+  createdAt: number;
+  type: string;
+  data: NotificationsMap[key];
+};
+
+export type Subscription = {
+  transfers: {
+    amount: number;
+    denom: string;
+    fromAddress: string;
+    toAddress: string;
+    blockHeight: number;
+  };
+};
+
+type UseNotificationsParameters = {
+  limit?: number;
+  page?: number;
+};
+
+export const notifier = createEventBus<NotificationsMap>();
+
+export function useNotifications(parameters: UseNotificationsParameters = {}) {
+  const { limit = 5, page = 1 } = parameters;
+
+  const { coins } = useConfig();
 
   const [__notifications__, setNotifications] = useStorage<
     { type: string; data: unknown; createdAt: number }[]
   >("app.notifications", { initialValue: [], version: 0.1 });
-  const pushNotification = useCallback(
-    (notification: { type: string; data: unknown; createdAt: number }) => {
-      setNotifications((prev) => [...prev, notification]);
-    },
-    [],
-  );
 
-  const hasNotifications = __notifications__.length > 0;
+  const totalNotifications = __notifications__.length;
+  const hasNotifications = totalNotifications > 0;
 
-  const notifications: Record<string, Notifications[]> = useMemo(
-    () =>
-      [...__notifications__]
-        .reverse()
-        .slice(0, maxNotifications || __notifications__.length)
-        .sort((a, b) => b.createdAt - a.createdAt)
-        .reduce((acc, notification) => {
-          const dateKey = isToday(notification.createdAt)
-            ? "Today"
-            : format(notification.createdAt, "MM/dd/yyyy");
+  const notifications: Record<string, Notifications[]> = useMemo(() => {
+    const current = (page - 1) * limit;
+    return [...__notifications__]
+      .reverse()
+      .slice(current, current + limit)
+      .sort((a, b) => b.createdAt - a.createdAt)
+      .reduce((acc, notification) => {
+        const dateKey = isToday(notification.createdAt)
+          ? "Today"
+          : format(notification.createdAt, "MM/dd/yyyy");
 
-          if (!acc[dateKey]) {
-            acc[dateKey] = [];
-          }
-          acc[dateKey].push(notification);
-          return acc;
-        }, Object.create({})),
-    [maxNotifications, __notifications__],
-  );
+        if (!acc[dateKey]) {
+          acc[dateKey] = [];
+        }
+        acc[dateKey].push(notification);
+        return acc;
+      }, Object.create({}));
+  }, [__notifications__, limit, page]);
+
+  const subscribe = useCallback((account: Account) => {
+    let client: GraphqlSubscriptionClient | undefined;
+    (async () => {
+      client = createClient({ url: GRAPHQL_URI });
+      const subscription = client.iterate({
+        query: `subscription($address: String) {
+              sentTransfers: transfers(fromAddress: $address) {
+                fromAddress
+                toAddress
+                blockHeight
+                amount
+                denom
+              }
+              receivedTransfers: transfers(toAddress: $address) {
+                fromAddress
+                toAddress
+                blockHeight
+                amount
+                denom
+              }
+            }`,
+        variables: { address: account?.address },
+      });
+      for await (const { data } of subscription) {
+        if (!data) continue;
+        if ("receivedTransfers" in data || "sentTransfers" in data) {
+          const isSent = "sentTransfers" in data;
+
+          const [transfer] = data[
+            isSent ? "sentTransfers" : "receivedTransfers"
+          ] as Subscription["transfers"][];
+          if (!transfer) continue;
+          const coin = coins[transfer.denom];
+          const notification = {
+            ...transfer,
+            type: isSent ? "sent" : "received",
+            coin,
+          } as NotificationsMap["transfer"];
+
+          notifier.publish("transfer", notification);
+          setNotifications((prev) => [
+            ...prev,
+            {
+              type: "transfer",
+              data: notification,
+              createdAt: Date.now(),
+            },
+          ]);
+        }
+      }
+    })();
+    return () => {
+      if (client) client.dispose();
+    };
+  }, []);
 
   return {
+    notifier,
+    subscribe,
     notifications,
-    pushNotification,
     hasNotifications,
+    totalNotifications,
   };
 }
