+++ ui/portal/web/src/pages/(app)/_app.notifications.lazy.tsx
@@ -1,19 +1,13 @@
 import { createLazyFileRoute } from "@tanstack/react-router";
 
-import {
-  IconButton,
-  IconChevronDown,
-  IconGear,
-  twMerge,
-  useMediaQuery,
-} from "@left-curve/applets-kit";
-import { NotificationsList } from "~/components/notifications/NotificationsList";
+import { IconButton, IconChevronDown, twMerge, useMediaQuery } from "@left-curve/applets-kit";
+import { Notifications } from "~/components/notifications/Notifications";
 
 export const Route = createLazyFileRoute("/(app)/_app/notifications")({
-  component: NotificationComponent,
+  component: NotificationApplet,
 });
 
-function NotificationComponent() {
+function NotificationApplet() {
   const { isMd } = useMediaQuery();
   return (
     <div className="w-full md:max-w-[50rem] mx-auto flex flex-col gap-4 p-4 pt-6 mb-16">
@@ -27,9 +21,11 @@ function NotificationComponent() {
           <span className="h3-bold">Notifcations</span>
         </h2>
       </div>
-      <div className={twMerge("bg-rice-25 w-full shadow-card-shadow rounded-xl")}>
-        <NotificationsList notifications={[]} />
-      </div>
+      <Notifications
+        maxNotifications={isMd ? 10 : 5}
+        withPagination
+        className="shadow-card-shadow"
+      />
     </div>
   );
 }
